+++ tests/unit/test_oneview_client.py
@@ -762,9 +762,9 @@ class OneViewClientTest(unittest.TestCase):
     def test_server_profile_templates_has_value(self):
         self.assertIsNotNone(self._oneview.server_profile_templates)
 
-    def test_lazy_loading_server_profile_templates(self):
-        server_profile_templates = self._oneview.server_profile_templates
-        self.assertEqual(server_profile_templates, self._oneview.server_profile_templates)
+    def test_server_profile_templates_return(self):
+        self.assertNotEqual(self._oneview.server_profile_templates,
+                            self._oneview.server_profile_templates)
 
     def test_server_profiles_has_right_type(self):
         self.assertIsInstance(self._oneview.server_profiles, ServerProfiles)